Fix credentials (Minor changes)
