+++ app/services/user_management_system/schemas.py
@@ -91,12 +91,3 @@ class AutomobileResponse(AutomobileBase):
 
 class AutomobilePut(BaseModel):
     pass
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    id: Optional[str] = None
