+++ app/shared/credentials.py
@@ -1,0 +1,44 @@
+from datetime import datetime, timedelta
+
+from fastapi.security import OAuth2PasswordBearer
+from jose import JWTError, jwt
+
+from app.core.config import settings
+
+
+"""
+There are 3 pieces of information that we need for
+the token,they are:
+- Secret_key
+- Algorithm
+- Expiration
+"""
+SECRET_KEY = settings.secret_key
+ALGORITHM = settings.algorithm
+ACCESS_TOKEN_EXPIRE_MINUTES = settings.access_token_expire_minutes
+
+
+def create_access_token(data: dict):
+    to_encode = data.copy()
+
+    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    to_encode.update({"exp": expire})
+
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+
+    return encoded_jwt
+
+
+def verify_access_token(token: OAuth2PasswordBearer, credentials_exception):
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        user_id = str(payload.get("user_id"))
+
+        if user_id is None:
+            raise credentials_exception
+
+        return user_id
+
+    except JWTError:
+        raise credentials_exception
+
